spacebox, openbox

+++ utils/checkArgs.js
@@ -7,7 +7,8 @@ const checkArgs = async (command, args, msg) => {
     if (command.min && !(command.min < args.length)) {
         let m = msg.reply(errors.invalidArgs);
         await delay(1000*5);
-        m.delete().catch(err => console.log(err));
+        if (m)
+            m.delete().catch(err => console.log(err));
         return false;
     }
